@@ -10,14 +10,15 @@
     "cheerio": "^1.0.0-rc.2",
     "cors": "^2.8.5",
     "express": "^4.16.4",
-    "netlify-lambda": "^1.0.3",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-scripts": "2.1.1",
-    "serverless-http": "^1.8.0",
-    "styled-components": "^4.1.1"
+    "netlify-lambda": "^1.4.3",
+    "react": "^16.8.4",
+    "react-dom": "^16.8.4",
+    "react-scripts": "2.1.8",
+    "serverless-http": "^1.9.1",
+    "styled-components": "^4.1.3"
   },
   "devDependencies": {
+    "eslint-plugin-react-hooks": "^1.5.1",
     "http-proxy-middleware": "^0.19.1",
     "npm-run-all": "^4.1.5"
   },
@@ -30,9 +31,6 @@
     "build:lambda": "netlify-lambda build src/lambda",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
   },
   "browserslist": [
     ">0.2%",
