@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from "react";
+import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { BASE_URL } from "constants.js";
 import styled from "styled-components";
@@ -19,75 +19,58 @@ const Referent = styled.a`
     }
 `;
 
-class Result extends Component {
-    state = {
-        song: null,
-        lyricsData: null
-    };
+const Result = ({ resultId }) => {
+    const [song, setSong] = useState(null);
+    const [lyricsData, setLyricsData] = useState(null);
 
-    scrapePage = path => {
+    const scrapePage = path => {
         axios.get(`${BASE_URL}scrape?path=${path}`).then(({ data }) => {
-            console.log(data);
-
-            this.setState({
-                lyricsData: data.lyricsData
-            });
-            
-            console.log(data.lyricsData);
-
-            //axios.get(`${BASE_URL}translate/`).then(({data}) => console.log(data.translation));
+            setLyricsData(data.lyricsData);
         });
     };
 
-    componentDidMount() {
-        const { resultId } = this.props;
+    useEffect(() => {
         axios.get(`${BASE_URL}proxy/songs/${resultId}`).then(result => {
-            this.setState({
-                song: result.data.response.song
-            });
-            this.scrapePage(result.data.response.song.path);
+            const song = result.data.response.song;
+            setSong(song);
+            scrapePage(song.path);
         });
-    }
+    }, [resultId]);
 
-    render() {
-        const { resultId } = this.props;
-        const { song, lyricsData } = this.state;
-
-        return (
-            <div>
-                <div>id: {resultId}</div>
-                {song ? (
-                    <div>
-                        <div>{song.primary_artist.name}</div>
-                    </div>
-                ) : null}
-                {lyricsData ? (
-                    <div>
-                        {lyricsData.map(({ text, referentId }) => {
-                            if (referentId) {
-                                return (
-                                    <Fragment>
-                                        <Referent
-                                            href={`/referent/${referentId}`}
-                                            data-referent-id={referentId}
-                                            dangerouslySetInnerHTML={{
-                                                __html: text.replace(
-                                                    /(?:\r\n|\r|\n)/g,
-                                                    "<br />"
-                                                )
-                                            }}
-                                        />
-                                        <br />
-                                    </Fragment>
-                                );
-                            }
-                            return <p>{text}</p>;
-                        })}
-                    </div>
-                ) : null}
-            </div>
-        );
-    }
-}
+    return (
+        <div>
+            <div>id: {resultId}</div>
+            {song ? (
+                <div>
+                    <div>{song.primary_artist.name}</div>
+                </div>
+            ) : null}
+            {lyricsData ? (
+                <div>
+                    {lyricsData.map(({ text, referentId }) => {
+                        if (referentId) {
+                            return (
+                                <>
+                                    <Referent
+                                        href={`/referent/${referentId}`}
+                                        data-referent-id={referentId}
+                                        dangerouslySetInnerHTML={{
+                                            __html: text.replace(
+                                                /(?:\r\n|\r|\n)/g,
+                                                "<br />"
+                                            )
+                                        }}
+                                    />
+                                    <br />
+                                </>
+                            );
+                        }
+                        return <p>{text}</p>;
+                    })}
+                </div>
+            ) : null}
+        </div>
+    );
+};
 
 export default Result;
